add processing-test with debug-output and some fixes

+++ src/core/clustering/cluster/nodeCluster.cpp
@@ -63,7 +63,7 @@ uint32_t NodeCluster::getNumberOfNodeBlocks() const
  * @brief NodeCluster::getNumberOfNode
  * @return
  */
-uint32_t NodeCluster::getNumberOfNodes() const
+uint16_t NodeCluster::getNumberOfNodes() const
 {
     return m_metaData.numberOfNodes;
 }
